+++ src/0_session.js
@@ -8,11 +8,13 @@ goog.provide('session');
 
  /**
  * @param {BranchStorage} storage
+ * @param {boolean=} first
  * @return {Object}
  */
-session.get = function(storage) {
+session.get = function(storage, first) {
+	var sessionString = first ? 'branch_session_first' : 'branch_session';
 	try {
-		return goog.json.parse(storage.get('branch_session')) || null;
+		return goog.json.parse(storage.get(sessionString, first)) || null;
 	}
 	catch (e) {
 		return null;
@@ -22,7 +24,9 @@ session.get = function(storage) {
  /**
  * @param {BranchStorage} storage
  * @param {Object} data
+ * @param {boolean=} first
  */
-session.set = function(storage, data) {
+session.set = function(storage, data, first) {
 	storage.set('branch_session', goog.json.serialize(data));
+	if (first) { storage.set('branch_session_first', goog.json.serialize(data), true); }
 }