+++ test/3_branch.js
@@ -173,10 +173,11 @@ describe('Branch', function() {
 
 		it('should store in session and call open with link_identifier from hash', function(done) {
 			if (testUtils.go("#r:12345")) {
-				var branch = initBranch(false), assert = testUtils.plan(2, done);
+				var branch = initBranch(false), assert = testUtils.plan(3, done);
 
 				branch.init(branch_sample_key, function(err, data) {
-					assert.equal(utils.readStore(branch._storage).click_id, '12345', 'click_id from link_identifier hash stored in session_id');
+					assert.equal('12345', JSON.parse(localStorage.getItem('branch_session_first')).click_id, 'hash session_id stored in local storage');
+					assert.equal('12345', utils.mobileUserAgent() ? '12345' : JSON.parse(sessionStorage.getItem('branch_session')).click_id, 'hash session_id saved in session storage');
 				});
 
 				requests[0].callback(null, browser_fingerprint_id);
@@ -197,7 +198,8 @@ describe('Branch', function() {
 				var branch = initBranch(false), assert = testUtils.plan(2, done);
 
 				branch.init(branch_sample_key, function(err, data) {
-					assert.equal(utils.readStore(branch._storage).click_id, '67890', 'click_id from link_identifier get param stored in session_id');
+					assert.equal('67890', JSON.parse(localStorage.getItem('branch_session_first')).click_id, 'get param match id stored in local storage');
+					assert.equal('67890', utils.mobileUserAgent() ? '67890' : JSON.parse(sessionStorage.getItem('branch_session')).click_id, 'get param match id saved in session storage');
 				});
 
 				requests[0].callback(null, browser_fingerprint_id);
@@ -327,6 +329,28 @@ describe('Branch', function() {
 			assert.equal(requests.length, 1, 'Request made');
 			requests[0].callback();
 			assert.deepEqual(requests[0].obj, testUtils.params({ }, [ 'browser_fingerprint_id' ]), 'All params sent');
+		});
+
+		it('should overwrite existing session_id, sessionLink, and identity_id\'s', function(done) {
+			var branch = initBranch(true), assert = testUtils.plan(6, done);
+			branch.logout(function(err) {
+				assert(!err, 'No error');
+			});
+
+			assert.equal(requests.length, 1, 'Request made');
+
+			var original_session_id = branch.session_id,
+				original_identity_id = branch.identity_id,
+				original_link = branch.sessionLink;
+			var new_session_id = "new_session",
+				new_identity_id = "new_id",
+				new_link = "new_link";
+
+			requests[0].callback(null, { "identity_id": new_identity_id, "session_id": new_session_id, "link": new_link });
+			assert.deepEqual(requests[0].obj, testUtils.params({ }, [ 'browser_fingerprint_id' ]), 'All params sent');
+			assert.equal(branch.session_id, new_session_id, "branch session was replaced");
+			assert.equal(branch.identity_id, new_identity_id, "branch identity was replaced");
+			assert.equal(branch.sessionLink, new_link, "link was replaced");
 		});
 	});